pyproject.toml
[project]
name = "movie-recommender"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "jinja2>=3.1.6",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.6",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.3",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "pip>=25.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "streamlit>=1.48.0",
    "uvicorn>=0.35.0",
]
